requirements.txt
langchain
langgraph
python-dotenv
pydantic
langchain-community
langchain-openai
ipykernel
diffusers
torch
transformers
pillow
streamlit
diffusers # Fro Running stable diffusion model 
torch # FRAMEWORK
transformers # NLP MODELS/ TOKENIZATION
pillow # SAVE/ MUNIPULATE IMAGE
